added test class and dependencies

+++ src/main/java/com/smash/food_punk/models/ProductComparator.java
@@ -12,8 +12,7 @@ public class ProductComparator implements Comparator<Product> {
 
 	@Override
 	public int compare(Product o1, Product o2) {
-		return Float.compare(o1.getPrice(), o2.getPrice());
+		return Double.compare(o1.getPrice(), o2.getPrice());
 	}
-
 
 }
